+++ src/main/java/se/liu/ida/hefquin/engine/queryplan/utils/LogicalToPhysicalPlanConverter.java
@@ -1,4 +1,4 @@
-package se.liu.ida.hefquin.engine.queryproc.impl.optimizer;
+package se.liu.ida.hefquin.engine.queryplan.utils;
 
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.physical.PhysicalPlan;
